Restructured the folders, and removed Identity Server.  Also removed Authorization.

+++ src/Budgeting.Domain/Entities/BudgetItem.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Budgeting.Domain.Entities
 {
@@ -9,7 +7,7 @@ namespace Budgeting.Domain.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Frequency { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
